make mlb class structure match the rest
