+++ lib/sportradar/api/baseball.rb
@@ -21,26 +21,6 @@ require_relative 'baseball/lineup'
 require_relative 'baseball/scoring'
 
 require_relative 'baseball/mlb'
-require_relative 'baseball/mlb/hierarchy'
+require_relative 'baseball/mlb/api'
 require_relative 'baseball/mlb/league'
 require_relative 'baseball/mlb/division'
-# require_relative 'baseball/mlb/season'
-# require_relative 'baseball/mlb/schedule'
-# require_relative 'baseball/mlb/game'
-# require_relative 'baseball/mlb/team'
-# require_relative 'baseball/mlb/player'
-# require_relative 'baseball/mlb/quarter'
-
-# require_relative 'baseball/ncaamb'
-# require_relative 'baseball/ncaamb/season'
-# require_relative 'baseball/ncaamb/schedule'
-# require_relative 'baseball/ncaamb/tournament'
-# require_relative 'baseball/ncaamb/round'
-# require_relative 'baseball/ncaamb/bracket'
-# require_relative 'baseball/ncaamb/game'
-# require_relative 'baseball/ncaamb/half'
-# require_relative 'baseball/ncaamb/hierarchy'
-# require_relative 'baseball/ncaamb/division'
-# require_relative 'baseball/ncaamb/conference'
-# require_relative 'baseball/ncaamb/team'
-# require_relative 'baseball/ncaamb/player'
