+++ lib/sportradar/api/baseball/game.rb
@@ -465,7 +465,7 @@ module Sportradar
         end
 
         def api
-          @api || Sportradar::Api::Baseball::Mlb.new
+          @api || Sportradar::Api::Baseball::Mlb::Api.new
         end
 
         def sim!
